# Conflicts:
#	src/main/java/it/alten/pokemonao/configuration/CommonConfiguration.java
#	src/main/java/it/alten/pokemonao/models/PokemonModel.java
#	src/main/java/it/alten/pokemonao/models/TypeModel.java

+++ src/main/java/it/alten/pokemonao/models/TypeModel.java
@@ -1,13 +1,10 @@
 package it.alten.pokemonao.models;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 @Data
 public class TypeModel {
-    private long id;
+    private Integer id;
     private String name;
     private String icon;
 }
